requirements.txt
click==7.1.2
falcon==3.0.1
falcon-cors==1.1.7
falcon_auth==1.1.0
google-api-core==1.26.3
google-auth==1.30.0
google-cloud-bigquery==2.15.0
google-cloud-bigquery-storage==2.4.0
google-cloud-core==1.6.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
grpcio==1.37.1
Logbook==1.5.3
numpy==1.19.5
pandas==1.1.5
pyarrow==4.0.0
pycparser==2.20
pydantic==1.8.1
pytz==2019.1
PyYAML==5.4.1
requests==2.25.1
ujson==4.0.2
uvicorn==0.13.4
